requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kaggle==1.7.4.5
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
